Refactor App Package Name
