+++ app/src/main/java/com/yalematta/trending/TrendingApplication.kt
@@ -1,4 +1,4 @@
-package com.yalematta.trendingrepos
+package com.yalematta.trending
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
