+++ app/src/main/java/com/yalematta/trending/ui/repos/adapter/ReposLoadStateAdapter.kt
@@ -1,4 +1,4 @@
-package com.yalematta.trendingrepos.ui.repos.adapter
+package com.yalematta.trending.ui.repos.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,7 +6,7 @@ import androidx.core.view.isVisible
 import androidx.paging.LoadState
 import androidx.paging.LoadStateAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.yalematta.trendingrepos.databinding.ReposLoadStateFooterBinding
+import com.yalematta.trending.databinding.ReposLoadStateFooterBinding
 
 class ReposLoadStateAdapter(private val retry: () -> Unit) : LoadStateAdapter<ReposLoadStateAdapter.LoadStateViewHolder>(){
