+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/reposFragment"
-        android:name="com.yalematta.trendingrepos.ui.repos.ReposFragment"
+        android:name="com.yalematta.trending.ui.repos.ReposFragment"
         android:label="fragment_repos"
         tools:layout="@layout/fragment_repos" >
         <action
@@ -20,11 +20,11 @@
     </fragment>
     <fragment
         android:id="@+id/detailsFragment"
-        android:name="com.yalematta.trendingrepos.ui.details.DetailsFragment"
+        android:name="com.yalematta.trending.ui.details.DetailsFragment"
         android:label="fragment_details"
         tools:layout="@layout/fragment_details" >
         <argument
             android:name="repo"
-            app:argType="com.yalematta.trendingrepos.data.model.Repo" />
+            app:argType="com.yalematta.trending.data.model.Repo" />
     </fragment>
 </navigation>
